requirements.txt
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
inflate64==1.0.0
multivolumefile==0.2.3
psutil==6.1.0
py7zr==0.22.0
pybcj==1.0.2
pycryptodomex==3.21.0
pyppmd==1.1.0
PyYAML==6.0.2
pyzstd==0.16.2
requests==2.32.3
soupsieve==2.6
texttable==1.7.0
urllib3==2.2.3
